Por ahora todo funciona. No cambies nada :bat:

+++ src/coord.cpp
@@ -3,15 +3,15 @@
 
 Coord::Coord() : x(0), y(0) { }
 
-Coord::Coord(int x, int y) : x(x), y(y) { }
+Coord::Coord(double x, double y) : x(x), y(y) { }
 
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 /*                          Getters and Setters                          */
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-int Coord::getx() { return x; }
+double Coord::getx() { return x; }
 
-int Coord::gety() { return y; }
+double Coord::gety() { return y; }
 
 void Coord::randomize(int size) {
 	x = rand() % size + 1; /* range from 1 to size */
@@ -19,6 +19,6 @@ void Coord::randomize(int size) {
 }
 
 
-void Coord::setx(int x) { this->x = x; }
+void Coord::setx(double x) { this->x = x; }
 
-void Coord::sety(int y) { this->y = y; }
+void Coord::sety(double y) { this->y = y; }
